+++ usi/src/main/java/daq/usi/ResponseHandler.java
@@ -1,0 +1,5 @@
+package daq.usi;
+
+public interface ResponseHandler<T> {
+  void receiveData(T data) throws Exception;
+}